Move FakeDatetimeFormatter into tests/www module
