+++ tests/datetime_service.py
@@ -3,7 +3,6 @@ from typing import Optional
 
 from arbeitszeit.datetime_service import DatetimeService
 from arbeitszeit.injector import singleton
-from arbeitszeit_web.formatters.datetime_formatter import DatetimeFormatter
 
 
 def datetime_utc(
@@ -43,17 +42,3 @@ class FakeDatetimeService(DatetimeService):
 
     def now_minus(self, delta: timedelta) -> datetime:
         return self.now() - delta
-
-
-class FakeDatetimeFormatter(DatetimeFormatter):
-    def format_datetime(
-        self,
-        date: datetime,
-        fmt: Optional[str] = None,
-    ) -> str:
-        if date.tzinfo is None:
-            date = date.replace(tzinfo=UTC)
-        date = date.astimezone(UTC)
-        if fmt is None:
-            fmt = "%d.%m.%Y %H:%M"
-        return date.strftime(fmt)
